Extract a new {Factory} class from {Builder}
Now {Factory} is the only object that knows how to construct all of the
other objects in this library. The {Builder} class accepts a {Factory}
instance and only needs to know how to use the dependencies.
