+++ lib/sub_diff.rb
@@ -6,6 +6,7 @@ require 'sub_diff/builder'
 require 'sub_diff/collection'
 require 'sub_diff/diff'
 require 'sub_diff/differ'
+require 'sub_diff/factory'
 require 'sub_diff/sub'
 require 'sub_diff/gsub'
 require 'sub_diff/version'
